WIP: add class BestFirstSearch
The class represents the algorithm for finding the path from
the source vertex to the destination vertex.
Principle (simple): select the next vertex on the principle
of least weight.

TODO: The class needs to implement methods for backtracking and
calculating the weight of the path, and then testing the algorithm